+++ routes/web.php
@@ -17,37 +17,41 @@ use App\Http\Controllers\ListingController;
 */
 
 // All listings
-Route::get('/', [ListingController::class, 'index']);
+Route::get('/', [ListingController::class, 'index'])->name('home');
 
 // Store Listing Form Data 
-Route::post('/listings' , [ListingController::class,'store']);
+Route::post('/listings' , [ListingController::class,'store'])->middleware('auth');
 
 // Show Create Form
-Route::get('/listings/create' , [ListingController::class,'create']);
+Route::get('/listings/create' , [ListingController::class,'create'])->middleware('auth');
+
+
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // Single listings
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 // Edit listing
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // Update listing
-Route::put('/listings/{listing}', [ListingController::class,'update']);
+Route::put('/listings/{listing}', [ListingController::class,'update'])->middleware('auth');
 
 // Show Register/Create Form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Create New User
-Route::post('/users', [UserController::class, 'store']);
+Route::post('/users', [UserController::class, 'store'])->middleware('guest');
 
 // Log Out
-Route::post('/logout', [UserController::class,'logout']);
+Route::post('/logout', [UserController::class,'logout'])->middleware('auth');
 
 // Show Login Form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Log in user
-Route::post('/users/authenticate', [UserController::class,'authenticate']);
+Route::post('/users/authenticate', [UserController::class,'authenticate'])->middleware('guest');
